fix add-image

+++ src/java/store/controllers/AddImageController.java
@@ -19,6 +19,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.Part;
 import org.apache.commons.io.FilenameUtils;
 import store.shopping.ProductDAO;
+import store.shopping.ProductDTO;
 import store.utils.VNCharacterUtils;
 
 /**
@@ -30,7 +31,7 @@ import store.utils.VNCharacterUtils;
 public class AddImageController extends HttpServlet {
 
     private static final String ERROR = "error.jsp";
-    private static final String SUCCESS = "MainController?action=ViewImages";
+    private static final String SUCCESS = "ViewImagesController";
 
     private File uploadFolder;
 
@@ -54,34 +55,39 @@ public class AddImageController extends HttpServlet {
         String url = ERROR;
         try {
             ProductDAO dao = new ProductDAO();
-            int productID = Integer.parseInt(request.getParameter("productID"));
-            String productName = request.getParameter("productName");
+            int productID = Integer.parseInt(request.getParameter("productID"));    
             String color = request.getParameter("color");
+            ProductDTO product = dao.getProductDetail(productID);
+            if (product.getColorImage().get(color).size() < 3) {
+                String productName = request.getParameter("productName");
+                int productColorID = dao.getProductColorID(productID, color);
+                color = VNCharacterUtils.removeAccent(color).toLowerCase().trim();
+                String uploadPath = getServletContext().getInitParameter("uploadFolder");
+                List<Part> images = request.getParts().stream().filter(part -> "image".equals(part.getName()) && part.getSize() > 0).collect(Collectors.toList()); // Retrieves <input type="file" name="files" multiple="true">
+                String fileNamePrefix = VNCharacterUtils.removeAccent(productName).trim().toLowerCase().replace(" ", "-");
+                for (Part image : images) {
+                    fileNamePrefix = fileNamePrefix + "-" + color;
+                    String fileName = FilenameUtils.getName(image.getSubmittedFileName());
+                    String fileNameSuffix = "." + FilenameUtils.getExtension(fileName);
+                    File file = File.createTempFile(fileNamePrefix, fileNameSuffix, uploadFolder);
+                    image.write(file.getAbsolutePath());
 
-            int productColorID = dao.getProductColorID(productID, color);
-            color = VNCharacterUtils.removeAccent(color).toLowerCase().trim();
-            String uploadPath = getServletContext().getInitParameter("uploadFolder");
-            List<Part> images = request.getParts().stream().filter(part -> "image".equals(part.getName()) && part.getSize() > 0).collect(Collectors.toList()); // Retrieves <input type="file" name="files" multiple="true">
-            String fileNamePrefix = VNCharacterUtils.removeAccent(productName).trim().toLowerCase().replace(" ", "-");
-            for (Part image : images) {
-                fileNamePrefix = fileNamePrefix + "-" + color;
-                String fileName = FilenameUtils.getName(image.getSubmittedFileName());
-                String fileNameSuffix = "." + FilenameUtils.getExtension(fileName);
-                File file = File.createTempFile(fileNamePrefix, fileNameSuffix, uploadFolder);
-                image.write(file.getAbsolutePath());
+                    boolean check = dao.addImage(productColorID, "/" + uploadPath + file.getName());
 
-                boolean check = dao.addImage(productColorID, "/" + uploadPath + file.getName());
-
-                if (check) {
-                    url = SUCCESS;
-                    request.setAttribute("MESSAGE", "Thêm ảnh thành công!");
-                } else {
-                    file.delete();
+                    if (check) {
+                        url = SUCCESS;
+                        request.setAttribute("MESSAGE", "Thêm ảnh thành công!");
+                    } else {
+                        file.delete();
+                    }
                 }
+            } else {
+                url = SUCCESS;
+                request.setAttribute("MESSAGE", "Số lượng ảnh quá giới hạn cho phép!");
             }
 
         } catch (IOException | NumberFormatException | SQLException | ServletException e) {
-            log("Error at ManagerShowProductDetailController: " + e.toString());
+            log("Error at AddImageController: " + e.toString());
         } finally {
             request.getRequestDispatcher(url).forward(request, response);
         }
